Refactor server/lib/gh-api into dirs for each base endpoint
