+++ src/server/services/validateTokenService.ts
@@ -1,5 +1,5 @@
 import { SHARED_ValidateTokenAPIResponse } from '@/models/shared'
-import { GH_API_getUser } from '@/server/lib/gh-api/user'
+import { GH_API_getUser } from '@/server/lib/gh-api/users'
 
 export const validateAccessToken = async (accessToken: string): Promise<SHARED_ValidateTokenAPIResponse> => {
     const { success, error } = await GH_API_getUser(accessToken)
